resume-screening-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9           # Postgres driver
python-multipart==0.0.9           # file uploads
resume-matcher==0.2               # ATS match score [17]
pdfminer.six==20231228            # PDF → text
python-docx==1.1.2                # DOCX → text
spacy
transformers==4.41.2
huggingface-hub==0.24.0
python-dotenv
scikit-learn==1.4.2